backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-socketio==5.11.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Celery and task queue
celery==5.3.6
redis==5.0.1
kombu==5.3.5
flower==2.0.1

# Langchain and LLM
langchain==0.1.4
langchain-community==0.0.16
langchain-google-genai==0.0.6
google-generativeai==0.3.2
langchain-core==0.1.14

# Vector store and embeddings
weaviate-client==4.4.1
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pdf2image==1.17.0
pytesseract==0.3.10
pypdf==4.0.1

# Text processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.2

# Monitoring and performance
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
